+++ stable_baselines3/common/vec_env/vec_monitor.py
@@ -23,6 +23,9 @@ class VecMonitor(VecEnvWrapper):
     :param info_keywords: extra information to log, from the information return of env.step()
     """
 
+    episode_returns: np.ndarray
+    episode_lengths: np.ndarray
+
     def __init__(
         self,
         venv: VecEnv,
@@ -49,8 +52,6 @@ class VecMonitor(VecEnvWrapper):
             )
 
         VecEnvWrapper.__init__(self, venv)
-        self.episode_returns = None
-        self.episode_lengths = None
         self.episode_count = 0
         self.t_start = time.time()
 
@@ -58,12 +59,12 @@ class VecMonitor(VecEnvWrapper):
         if hasattr(venv, "spec") and venv.spec is not None:
             env_id = venv.spec.id
 
+        self.results_writer: Optional[ResultsWriter] = None
         if filename:
             self.results_writer = ResultsWriter(
-                filename, header={"t_start": self.t_start, "env_id": env_id}, extra_keys=info_keywords
+                filename, header={"t_start": self.t_start, "env_id": str(env_id)}, extra_keys=info_keywords
             )
-        else:
-            self.results_writer = None
+
         self.info_keywords = info_keywords
 
     def reset(self) -> VecEnvObs: